Move integration test utility functions to their own module.
