+++ tests/integration/test_infrahubctl.py
@@ -2,12 +2,9 @@
 
 import json
 import os
-import re
 import shutil
 import tempfile
-from contextlib import contextmanager
 from pathlib import Path
-from typing import Generator
 
 import pytest
 from git import Repo
@@ -16,26 +13,12 @@ from typer.testing import Any, CliRunner
 
 from infrahub_sdk.ctl.cli_commands import app
 
+from .utils import change_directory, strip_color
+
 runner = CliRunner()
 
 
 FIXTURE_BASE_DIR = Path(Path(os.path.abspath(__file__)).parent / ".." / "fixtures" / "integration" / "test_infrahubctl")
-
-
-@contextmanager
-def change_directory(new_directory: str) -> Generator[None, None, None]:
-    """Helper function used to change directories in a with block."""
-    # Save the current working directory
-    original_directory = os.getcwd()
-
-    # Change to the new directory
-    try:
-        os.chdir(new_directory)
-        yield  # Yield control back to the with block
-
-    finally:
-        # Change back to the original directory
-        os.chdir(original_directory)
 
 
 def read_fixture(file_name: str, fixture_subdir: str = ".") -> Any:
@@ -44,11 +27,6 @@ def read_fixture(file_name: str, fixture_subdir: str = ".") -> Any:
         fixture_contents = fhd.read()
 
     return fixture_contents
-
-
-def strip_color(text: str) -> str:
-    ansi_escape = re.compile(r"\x1B[@-_][0-?]*[ -/]*[@-~]")
-    return ansi_escape.sub("", text)
 
 
 @pytest.fixture
